test: 포인트 차감/충전 동시성 테스트
